payroll attendance  varibules bug fix(apddcf )
git-svn-id: svn+ssh://svn.vasista.in/srv/svn/repos/kmf/trunk@3001 40d85018-ee8e-4105-aa2c-066e80ea2389

+++ applications/humanres/src/in/vasista/vbiz/humanres/PayrollService.java
@@ -2885,7 +2885,8 @@ public class PayrollService {
 						String variableName = (String)tempEntry.getKey();
 						if(UtilValidate.isNotEmpty(tempEntry.getValue()) && tempEntry.getValue() instanceof BigDecimal){
 							result.put(variableName, 0.0);
-							result.put(variableName, (BigDecimal)tempEntry.getValue());
+							//result.put(variableName, (BigDecimal)tempEntry.getValue());
+							result.put(variableName, ((BigDecimal)tempEntry.getValue()).doubleValue());
 						}
 						
 					}
